- Endringer på Bilsøket og verdier. - Endret på Footer - Lagt til border custom colors - Fikset skrivefeil - Lagt til landing for BilNokler
