+++ app/bil-nokler/page.tsx
@@ -6,7 +6,7 @@ export default function BilNokler() {
         <div>
             <BilNoklerLanding />
             <div id="services">
-                {/* <CarKeySearch /> */}
+                <CarKeySearch />
             </div>
         </div>
     )
