+++ app/home/components/HomeBreak.tsx
@@ -5,7 +5,7 @@ const HomeBreak = () => {
     <div className='bg-custom-brown text-white'>
       <div className='max-w-7xl mx-auto flex flex-col py-16'>
         <h2 className='text-7xl uppercase font-semibold'>Reparer i dag!</h2>
-        <p className='font-sans mt-4 max-w-xl'>Kom innom oss eller kontakt oss for en gratis befaring. Noen av våre tjenester kan ta mer enne 1 dag, men de fleste kan gjøres menst du venter.</p>
+        <p className='font-sans mt-4 max-w-xl'>Kom innom oss eller kontakt oss for en gratis befaring. Noen av våre tjenester kan ta mer enn 1 dag, men de fleste kan gjøres menst du venter.</p>
       </div>
     </div>
   )
